unified Titan Graph Output Writer/Format
